Add files via upload
Started work on accessing a user's calendar. Currently can run locally and print to a console. 

Probably a good idea to change our api key's and client_secret stuff in the implementation. Maybe want to work on the js implementation more. 

You do need client_secret.json to run this. 

Some of this code was obtained by google's tutorials.