Update application.py
Chris added quite a bit for this one, but i dont think he committed those changes yet. Most of this commit is not me - but wanted to make sure that my change got logged so that we don't revert back to a non-working version if we do any reverts. 

I just made a change at line 1766 to make sure that access is used properly (as an integer) instead of (i believe it was unicode which i believe was being treated like a string. )